Move CustomDropdownView into shared file
