+++ src/elements/demo-image/DemoImageElementForm.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { CustomDropdown } from "../../editorial-source-components/CustomDropdown";
 import { Label } from "../../editorial-source-components/Label";
-import type { Option } from "../../plugin/fieldViews/DropdownFieldView";
+import { CustomDropdownView } from "../../plugin/fieldViews/CustomDropdownView";
 import type { FieldNameToValueMap } from "../../plugin/fieldViews/helpers";
 import type {
   CustomFieldViewSpec,
@@ -97,28 +96,5 @@ const ImageView = ({ fieldViewSpec }: ImageViewProps) => {
         </button>
       )}
     </div>
-  );
-};
-
-type CustomDropdownViewProps = {
-  fieldViewSpec: CustomFieldViewSpec<string, Array<Option<string>>>;
-};
-
-const CustomDropdownView = ({ fieldViewSpec }: CustomDropdownViewProps) => {
-  const [selectedElement, setSelectFieldsRef] = useCustomFieldViewState(
-    fieldViewSpec
-  );
-  return (
-    <CustomDropdown
-      options={fieldViewSpec.fieldSpec.props}
-      selected={selectedElement}
-      label={fieldViewSpec.name}
-      onChange={(event) => {
-        if (setSelectFieldsRef.current) {
-          setSelectFieldsRef.current(event.target.value);
-        }
-      }}
-      dataCy={getFieldViewTestId(fieldViewSpec.name)}
-    />
   );
 };
