+++ src/elements/embed/EmbedForm.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { CustomCheckbox } from "../../editorial-source-components/CustomCheckbox";
-import { CustomDropdown } from "../../editorial-source-components/CustomDropdown";
 import { Label } from "../../editorial-source-components/Label";
-import type { Option } from "../../plugin/fieldViews/DropdownFieldView";
+import { CustomDropdownView } from "../../plugin/fieldViews/CustomDropdownView";
 import type { FieldNameToValueMap } from "../../plugin/fieldViews/helpers";
 import type {
   CustomFieldViewSpec,
@@ -43,29 +42,6 @@ export const EmbedElementForm: React.FunctionComponent<Props> = ({
     <pre>{JSON.stringify(fields)}</pre>
   </div>
 );
-
-type CustomDropdownViewProps = {
-  fieldViewSpec: CustomFieldViewSpec<string, Array<Option<string>>>;
-};
-
-const CustomDropdownView = ({ fieldViewSpec }: CustomDropdownViewProps) => {
-  const [selectedElement, setSelectFieldsRef] = useCustomFieldViewState(
-    fieldViewSpec
-  );
-  return (
-    <CustomDropdown
-      options={fieldViewSpec.fieldSpec.props}
-      selected={selectedElement}
-      label={fieldViewSpec.name}
-      onChange={(event) => {
-        if (setSelectFieldsRef.current) {
-          setSelectFieldsRef.current(event.target.value);
-        }
-      }}
-      dataCy={getFieldViewTestId(fieldViewSpec.name)}
-    />
-  );
-};
 
 type CustomCheckboxViewProps = {
   fieldViewSpec: CustomFieldViewSpec<boolean, boolean>;
