+++ src/plugin/fieldViews/CustomDropdownView.tsx
@@ -1,0 +1,30 @@
+import { CustomDropdown } from "../../editorial-source-components/CustomDropdown";
+import { getFieldViewTestId } from "../../renderers/react/FieldView";
+import { useCustomFieldViewState } from "../../renderers/react/useCustomFieldViewState";
+import type { CustomFieldViewSpec } from "../types/Element";
+import type { Option } from "./DropdownFieldView";
+
+type CustomDropdownViewProps = {
+  fieldViewSpec: CustomFieldViewSpec<string, Array<Option<string>>>;
+};
+
+export const CustomDropdownView = ({
+  fieldViewSpec,
+}: CustomDropdownViewProps) => {
+  const [selectedElement, setSelectFieldsRef] = useCustomFieldViewState(
+    fieldViewSpec
+  );
+  return (
+    <CustomDropdown
+      options={fieldViewSpec.fieldSpec.props}
+      selected={selectedElement}
+      label={fieldViewSpec.name}
+      onChange={(event) => {
+        if (setSelectFieldsRef.current) {
+          setSelectFieldsRef.current(event.target.value);
+        }
+      }}
+      dataCy={getFieldViewTestId(fieldViewSpec.name)}
+    />
+  );
+};
